Added simple rows based on the column options
Thinking through some things:
Efficient way to pass around the options and source.
What sort of form and info the options need.
Keep the possibility of having any form of source and handle them all the same way. No exceptions or differences.
Way of keeping the column widths the same for all rows but still be dynamic.
Need to move the css into global file so that they all can be changed by devs.
The column classes could hold a view and editor for all cells in the column.